// src/app/libraries/lily/theme/models/sub-themes/buttons/buttons.ts
/**
 * Describes style properties for Lily buttons.
 */
export interface LilyButtonsTheme {
  /**
   * Background color of the button.
   */
  lilyButtonBackgroundColor: string;

  /**
   * Border color of the button.
   */
  lilyButtonBorderColor: string;

  /**
   * Font color of the button.
   */
  lilyButtonFontColor: string;

  /**
   * Background color of the button when focused.
   */
  lilyButtonFocusBackgroundColor: string;

  /**
   * Border color of the button when focused.
   */
  lilyButtonFocusBorderColor: string;

  /**
   * Font color of the button when focused.
   */
  lilyButtonFocusFontColor: string;

  /**
   * Background color of the button when disabled.
   */
  lilyButtonDisabledBackgroundColor: string;

  /**
   * Border color of the button when disabled.
   */
  lilyButtonDisabledBorderColor: string;

  /**
   * Font color of the button when disabled.
   */
  lilyButtonDisabledFontColor: string;

  /**
   * Background color of the button when hovered.
   */
  lilyButtonHoverBackgroundColor: string;

  /**
   * Border color of the button when hovered.
   */
  lilyButtonHoverBorderColor: string;

  /**
   * Font color of the button when hovered.
   */
  lilyButtonHoverFontColor: string;
}